# reborn_rentals_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API – Proyecto X",
        "description": "Documentación de la API con autenticación JWT (Bearer).",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Listar categorías (público)",
                "description": "GET /api/categories\nPúblico: lista todas las categorías.",
                "operationId": "8e17a9627dbed245cf032f867569cf11",
                "responses": {
                    "200": {
                        "description": "Listado de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existen categorías registradas aún"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Crear categoría (admin)",
                "description": "POST /api/categories\nProtegido: solo admin puede crear.",
                "operationId": "2140f330d926c3e118cab9391dc4ff42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Electrodomésticos"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Linea blanca y cocina",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Ver una categoría (público)",
                "description": "GET /api/category/{id}\nPúblico: cualquier usuario puede ver detalles.",
                "operationId": "5135d626104e86ea9910eb51763b8d4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Actualizar categoría (admin)",
                "description": "PUT /api/category/{id}\nProtegido: solo admin puede actualizar.",
                "operationId": "6aea2996a1d8d5708b6b26adca6b0f86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Electrónica y audio"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Audio, video y gadgets",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Eliminar categoría (admin)",
                "description": "DELETE /api/category/{id}\nProtegido: solo admin puede eliminar.",
                "operationId": "19bebec474ebd26750dd56af34eac8fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminada"
                    },
                    "200": {
                        "description": "Eliminada (mensaje)"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electrónica"
                    },
                    "description": {
                        "type": "string",
                        "example": "Gadgets y dispositivos",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envia el token como: Bearer {tu_token_jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}